event/core-poll: a crappy hack for Microsoft's broken poll()
Whoever was in charge of implementing poll() on Windows failed to test
it properly, and now the lazy bastards won't even fix it because someone
might rely on the bugged behavior.

What this hack does is create a background thread specifically to wait
on an FD_CONNECT event with a non-zero error code. Error callback is then
queued to the main loop. If WSAPoll() returns first with a real event
(e.g. connection actually succeeds), the wait thread is cancelled.

Due to how this workaround is implemented, notifications on connection
failure are edge-triggered, that is, they're only delivered once.

See also: https://daniel.haxx.se/blog/2012/10/10/wsapoll-is-broken/