stream/hwdev/bda: "fix" universal provider memory leak
When first added to a graph, the universal NP registers some sort of
PBDA event service with the graph (SID_ESEventService), and then doesn't
fully release it when RemoveFilter is called. Registering a dummy object
under the event service's SID prevents the NP from querying
IESEventService, thereby avoiding the leak.

Legacy providers are not affected by this bug.